pallet_domains

Module bundle_storage_fund

Bundle storage fund

Structs

StorageFundRedeemPrice

Enums

AccountType

The type of system account being created.

Error

Bundle storage fund specific errors

Constants

STORAGE_FEE_RESERVE

The proportion of staking fund reserved for the bundle storage fee

Functions

charge_bundle_storage_fee

Charge the bundle storage fee from the operator’s bundle storage fund

deposit_reserve_for_storage_fund

Split the new deposit into 2 parts: the staking deposit and the storage fee deposit, add the storage fee deposit to the bundle storage fund.

refund_storage_fee

Refund the paid bundle storage fee of a particular domain block back to the operator, the amount to refund to a particular operator is determined by the total storage fee collected from the domain user and the percentage of bundle storage that the operator have submitted for the domain block.

storage_fund_account

Return the bundle storage fund account of the given operator.

storage_fund_redeem_price

Return the bundle storage fund redeem price

total_balance

Return the total balance of the bundle storage fund the given operator_id

transfer_all_to_treasury

Transfer all of the balance of the bundle storage fund to the treasury

withdraw_and_hold

Transfer the given withdraw_amount of balance from the bundle storage fund to the given dest_account and hold on the dest_account

withdraw_to

Transfer the given withdraw_amount of balance from the bundle storage fund to the given dest_account