pallet_domains
Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

deregister_operator

Warning: Doc-Only

force_staking_epoch_transition

Force staking epoch transition for a given domain

freeze_domain

Freezes a given domain. A frozen domain does not accept new bundles but accepts fraud proofs.

instantiate_domain

Warning: Doc-Only

nominate_operator

Warning: Doc-Only

prune_domain_execution_receipt

Prunes a given execution receipt for given frozen domain. This call assumes the execution receipt to be bad and implicitly trusts Sudo to do necessary validation of the ER before dispatching this call.

register_domain_runtime

Warning: Doc-Only

register_operator

Warning: Doc-Only

send_domain_sudo_call

Submit a domain sudo call.

set_permissioned_action_allowed_by

Update permissioned action allowed by storage by Sudo.

submit_bundle

Warning: Doc-Only

submit_fraud_proof

Warning: Doc-Only

submit_receipt

Warning: Doc-Only

transfer_treasury_funds

Transfer funds from treasury to given account

unfreeze_domain

Unfreezes a frozen domain.

unlock_funds

Unlocks the first withdrawal given the unlocking period is complete. Even if rest of the withdrawals are out of unlocking period, nominator should call this extrinsic to unlock each withdrawal

unlock_nominator

Unlocks the nominator under given operator given the unlocking period is complete. A nominator can initiate their unlock given operator is already deregistered.

update_domain_operator_allow_list

Extrinsic to update domain’s operator allow list. Note:

upgrade_domain_runtime

Warning: Doc-Only

withdraw_stake

Warning: Doc-Only