Function pallet_domains::migrations::migrate_domain_object::get_domain_object

pub fn get_domain_object<T: Config>(
    domain_id: DomainId
) -> Option<DomainObject<BlockNumberFor<T>, ReceiptHashFor<T>, <T as Config>::AccountId>>

Whenever there is a migration to specific storage and that storage item is used as part of the runtimeAPI, decoding such storage item fails Here is an example: If runtime upgrade happened on Block #100, migration is not done yet But since the :code already holds the new runtime, runtime api will use the new Runtime when queried at Block #100. This essentially causes a Decode to fail and cause State Corrupt error.

Once the migration is done at block #101, this runtime api works as expected. To adjust to the case for Block #100 specifically, we try to decode with latest type and then fall back to old type.