pallet_domains

Module staking

Staking for domains

Structs

DomainEpoch

Unique epoch identifier across all domains. A combination of Domain and its epoch.

NewDeposit

Operator

Type that represents an operator details.

OperatorConfig

OperatorDeregisteredInfo

SharePrice

A share price is parts per billion of shares/ai3. Note: Shares must always be equal to or lower than ai3, and both shares and ai3 can’t be zero.

StakingSummary

Enums

Error

OperatorStatus

Type that represents an operator status.

Functions

do_deactivate_operator

Deactivates a given operator. Operator status is marked as Deactivated with epoch_index after which they can reactivate back into operator set. Their stake is removed from the total domain stake since they will not be producing bundles anymore until re-registration.

do_deregister_operator

Deregisters a given operator who is either registered or deactivated. Operator is removed from the next operator set.

do_mark_invalid_bundle_authors

Mark all the invalid bundle authors from this ER and remove them from operator set. NOTE: any changes to this must be reflected in the fuzz_utils’ equivalent

do_mark_operators_as_slashed

Freezes the slashed operators and moves the operator to be removed once the domain they are operating finishes the epoch.

do_nominate_operator

do_reactivate_operator

Reactivate a given deactivated operator if the activation delay in epochs has passed. The operator is added to next operator set and will be able to produce bundles from next epoch.

do_register_operator

do_reward_operators

Distribute the reward to the operators equally and drop any dust to treasury.

do_unlock_funds

Unlocks any withdraws that are ready to be unlocked.

do_unlock_nominator

Unlocks an already de-registered operator’s nominator given unlock wait period is complete.

do_unmark_invalid_bundle_authors

Unmark all the invalid bundle authors from this ER that were marked invalid. Assumed the ER is invalid and add the marked operators as registered and add them back to next operator set. NOTE: any changes to this must be reflected in the fuzz_utils’ equivalent

do_withdraw_stake

Withdraw some or all of the stake, using an amount of shares.

mark_invalid_bundle_author

unmark_invalid_bundle_author