Struct pallet_domains::Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn submit_bundle( origin: OriginFor<T>, opaque_bundle: OpaqueBundleOf<T> ) -> DispatchResultWithPostInfo

pub fn submit_fraud_proof( origin: OriginFor<T>, fraud_proof: Box<FraudProof<BlockNumberFor<T>, T::Hash, T::DomainHeader>> ) -> DispatchResult

pub fn register_domain_runtime( origin: OriginFor<T>, runtime_name: String, runtime_type: RuntimeType, raw_genesis_storage: Vec<u8> ) -> DispatchResult

pub fn upgrade_domain_runtime( origin: OriginFor<T>, runtime_id: RuntimeId, raw_genesis_storage: Vec<u8> ) -> DispatchResult

pub fn register_operator( origin: OriginFor<T>, domain_id: DomainId, amount: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance, config: OperatorConfig<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance> ) -> DispatchResult

pub fn nominate_operator( origin: OriginFor<T>, operator_id: OperatorId, amount: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance ) -> DispatchResult

pub fn instantiate_domain( origin: OriginFor<T>, domain_config: DomainConfig<T::AccountId, <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance> ) -> DispatchResult

pub fn deregister_operator( origin: OriginFor<T>, operator_id: OperatorId ) -> DispatchResult

pub fn withdraw_stake( origin: OriginFor<T>, operator_id: OperatorId, shares: T::Share ) -> DispatchResult

pub fn unlock_funds( origin: OriginFor<T>, operator_id: OperatorId ) -> DispatchResult

Unlocks the first withdrawal given the unlocking period is complete. Even if rest of the withdrawals are out of unlocking period, nominator should call this extrinsic to unlock each withdrawal

pub fn unlock_operator( origin: OriginFor<T>, operator_id: OperatorId ) -> DispatchResult

Unlocks the operator given the unlocking period is complete. Anyone can initiate the operator unlock.

pub fn update_domain_operator_allow_list( origin: OriginFor<T>, domain_id: DomainId, operator_allow_list: OperatorAllowList<T::AccountId> ) -> DispatchResult

Extrinsic to update domain’s operator allow list. Note:

If the previous allowed list is set to specific operators and new allow list is set to Anyone, then domain will become permissioned to open for all operators.

If the previous allowed list is set to Anyone or specific operators and the new allow list is set to specific operators, then all the registered not allowed operators will continue to operate until they de-register themselves.

pub fn force_staking_epoch_transition( origin: OriginFor<T>, domain_id: DomainId ) -> DispatchResult

Force staking epoch transition for a given domain

impl<T: Config> Pallet<T>

pub fn operator_signing_key<KArg>(k: KArg) -> Option<OperatorId>
where KArg: EncodeLike<OperatorPublicKey>,

An auto-generated getter for OperatorSigningKey.

impl<T: Config> Pallet<T>

pub fn domain_staking_summary<KArg>( k: KArg ) -> Option<StakingSummary<OperatorId, <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance>>
where KArg: EncodeLike<DomainId>,

An auto-generated getter for DomainStakingSummary.

impl<T: Config> Pallet<T>

pub fn consensus_block_info<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<T::Hash>
where KArg1: EncodeLike<DomainId>, KArg2: EncodeLike<BlockNumberFor<T>>,

An auto-generated getter for ConsensusBlockHash.

impl<T: Config> Pallet<T>

pub fn latest_submitted_er<KArg>(k: KArg) -> DomainBlockNumberFor<T>
where KArg: EncodeLike<(DomainId, OperatorId)>,

An auto-generated getter for LatestSubmittedER.

impl<T: Config> Pallet<T>

pub fn successful_bundles(domain_id: DomainId) -> Vec<H256>

pub fn successful_fraud_proofs(domain_id: DomainId) -> Vec<T::DomainHash>

pub fn domain_runtime_code(domain_id: DomainId) -> Option<Vec<u8>>

pub fn domain_best_number( domain_id: DomainId ) -> Option<DomainBlockNumberFor<T>>

pub fn runtime_id(domain_id: DomainId) -> Option<RuntimeId>

pub fn domain_instance_data( domain_id: DomainId ) -> Option<(DomainInstanceData, BlockNumberFor<T>)>

pub fn genesis_state_root(domain_id: DomainId) -> Option<H256>

pub fn domain_tx_range(domain_id: DomainId) -> U256

Returns the tx range for the domain.

pub fn bundle_producer_election_params( domain_id: DomainId ) -> Option<BundleProducerElectionParams<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance>>

pub fn operator( operator_id: OperatorId ) -> Option<(OperatorPublicKey, <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance)>

pub fn head_receipt_number(domain_id: DomainId) -> DomainBlockNumberFor<T>

Returns the best execution chain number.

pub fn oldest_unconfirmed_receipt_number( domain_id: DomainId ) -> Option<DomainBlockNumberFor<T>>

Returns the block number of the oldest existing unconfirmed execution receipt, return None means there is no unconfirmed ER exist or submitted yet.

pub fn latest_confirmed_domain_block_number( domain_id: DomainId ) -> DomainBlockNumberFor<T>

Returns the latest confirmed domain block number for a given domain Zero block is always a default confirmed block.

pub fn latest_confirmed_domain_block( domain_id: DomainId ) -> Option<(DomainBlockNumberFor<T>, T::DomainHash)>

pub fn domain_block_limit(domain_id: DomainId) -> Option<DomainBlockLimit>

Returns the domain block limit of the given domain.

pub fn domain_bundle_limit( domain_id: DomainId ) -> Result<Option<DomainBundleLimit>, DomainRegistryError>

Returns the domain bundle limit of the given domain

pub fn non_empty_er_exists(domain_id: DomainId) -> bool

Returns if there are any ERs in the challenge period that have non empty extrinsics. Note that Genesis ER is also considered special and hence non empty

pub fn extrinsics_shuffling_seed() -> T::Hash

pub fn execution_receipt( receipt_hash: ReceiptHashFor<T> ) -> Option<ExecutionReceiptOf<T>>

pub fn receipt_hash( domain_id: DomainId, domain_number: DomainBlockNumberFor<T> ) -> Option<ReceiptHashFor<T>>

pub fn confirmed_domain_block_storage_key(domain_id: DomainId) -> Vec<u8>

pub fn is_bad_er_pending_to_prune( domain_id: DomainId, receipt_number: DomainBlockNumberFor<T> ) -> bool

pub fn is_operator_pending_to_slash( domain_id: DomainId, operator_id: OperatorId ) -> bool

impl<T> Pallet<T>
where T: Config + SendTransactionTypes<Call<T>>,

pub fn submit_bundle_unsigned(opaque_bundle: OpaqueBundleOf<T>)

Submits an unsigned extrinsic Call::submit_bundle.

pub fn submit_fraud_proof_unsigned( fraud_proof: FraudProof<BlockNumberFor<T>, T::Hash, T::DomainHeader> )

Submits an unsigned extrinsic Call::submit_fraud_proof.