pallet_domains

Module tests

Structs

BlockAuthoringDelay

BlockReward

BlockSlotCount

BlockTreePruningDepth

BundleLongevity

ConfirmationDepthK

CurrentBundleAndExecutionReceiptVersion

DomainChainByteFee

DomainInstantiationDeposit

DomainTxRangeAdjustmentInterval

DomainsPalletId

DummyBlockSlot

DummyStorageFee

EraDuration

ExistentialDeposit

ExpectedVotesPerBlock

ExtrinsicStorageFees

HoldIdentifierWrapper

InitialDomainTxRange

InitialSolutionRange

MaxDomainBlockSize

MaxDomainBlockWeight

MaxDomainNameLength

MaxInitialDomainAccounts

MaxPendingStakingOperation

MaximumReceiptDrift

MinInitialDomainAccountBalance

MinNominatorStake

MinOperatorStake

MinSectorLifetime

MockBundleAndExecutionReceiptVersion

MockDomainsTransfersTracker

MockRandomness

OperatorActivationDelayInEpochs

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PotEntropyInjectionDelay

PotEntropyInjectionInterval

PotEntropyInjectionLookbackDepth

RecentHistoryFraction

RecentSegments

RecordSize

ReplicationFactor

ReportLongevity

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

ShouldAdjustSolutionRange

SlotProbability

StakeEpochDuration

StakeWithdrawalLockingPeriod

Test

TreasuryAccount

WithdrawalLimit

Enums

MockBundleVersion

MockExecutionReceiptVersion

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

DOMAIN_ID

INITIAL_SOLUTION_RANGE

Functions

new_test_ext

new_test_ext_with_extensions

Type Aliases

AccountId

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balances

BalancesConfig

BlockFees

DomainExecutive

Domains

DomainsConfig

MockVersionStore

Subspace

SubspaceConfig

System

SystemConfig

Timestamp