pallet_domains

Type Alias BlockTreeNodeFor

pub type BlockTreeNodeFor<T> = BlockTreeNode<BlockNumberFor<T>, <T as Config>::Hash, DomainBlockNumberFor<T>, <T as Config>::DomainHash, <T as Config>::Balance>;

Aliased Type

struct BlockTreeNodeFor<T> {
    pub execution_receipt: ExecutionReceipt<<<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number, <T as Config>::Hash, <<T as Config>::DomainHeader as Header>::Number, <T as Config>::DomainHash, <T as Config>::Balance>,
    pub operator_ids: Vec<u64>,
}

Fields
execution_receipt: ExecutionReceipt<<<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number, <T as Config>::Hash, <<T as Config>::DomainHeader as Header>::Number, <T as Config>::DomainHash, <T as Config>::Balance>

The full ER for this block.

operator_ids: Vec<u64>

A set of all operators who have committed to this ER within a bundle. Used to determine who to slash if a fraudulent branch of the block_tree is pruned.

NOTE: there may be duplicated operator id as an operator can submit multiple bundles with the same head receipt to a consensus block.