Type Alias pallet_domains::ConsensusBlockHash

pub type ConsensusBlockHash<T: Config> = StorageDoubleMap<_GeneratedPrefixForStorageConsensusBlockHash<T>, Identity, DomainId, Identity, BlockNumberFor<T>, T::Hash, OptionQuery>;

The consensus block hash used to verify ER, only store the consensus block hash for a domain if that consensus block contains bundle of the domain, the hash will be pruned when the ER that point to the consensus block is pruned.

TODO: this storage is unbounded in some cases, see https://github.com/subspace/subspace/issues/1673 for more details, this will be fixed once https://github.com/subspace/subspace/issues/1731 is implemented.

Storage type is [StorageDoubleMap] with key1 type DomainId, key2 type BlockNumberFor < T > and value type T :: Hash.

Aliased Type

struct ConsensusBlockHash<T: Config>(/* private fields */);