Type Alias pallet_domains::ExecutionInbox

pub type ExecutionInbox<T: Config> = StorageNMap<_GeneratedPrefixForStorageExecutionInbox<T>, (NMapKey<Identity, DomainId>, NMapKey<Identity, DomainBlockNumberFor<T>>, NMapKey<Identity, BlockNumberFor<T>>), Vec<BundleDigest<T::DomainHash>>, ValueQuery>;

A set of BundleDigest from all bundles that successfully submitted to the consensus block, these bundles will be used to construct the domain block and ExecutionInbox is used to:

Ensure subsequent ERs of that domain block include all pre-validated extrinsic bundles

Index the InboxedBundleAuthor and pruned its value when the corresponding ExecutionInbox is pruned

Storage type is [StorageNMap] with keys type (DomainId, DomainBlockNumberFor < T >, BlockNumberFor < T >) and value type Vec < BundleDigest < T :: DomainHash > >.

Aliased Type

struct ExecutionInbox<T: Config>(/* private fields */);