Type Alias pallet_domains::ExecutionReceiptOf

pub type ExecutionReceiptOf<T> = ExecutionReceipt<BlockNumberFor<T>, <T as Config>::Hash, DomainBlockNumberFor<T>, <T as Config>::DomainHash, <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance>;

Aliased Type

struct ExecutionReceiptOf<T> {
    pub domain_block_number: <<T as Config>::DomainHeader as Header>::Number,
    pub domain_block_hash: <T as Config>::DomainHash,
    pub domain_block_extrinsic_root: <T as Config>::DomainHash,
    pub parent_domain_block_receipt_hash: <T as Config>::DomainHash,
    pub consensus_block_number: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
    pub consensus_block_hash: <T as Config>::Hash,
    pub inboxed_bundles: Vec<InboxedBundle<<T as Config>::DomainHash>, Global>,
    pub final_state_root: <T as Config>::DomainHash,
    pub execution_trace: Vec<<T as Config>::DomainHash, Global>,
    pub execution_trace_root: H256,
    pub total_rewards: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
}

Fields
domain_block_number: <<T as Config>::DomainHeader as Header>::Number

The index of the current domain block that forms the basis of this ER.

domain_block_hash: <T as Config>::DomainHash

The block hash corresponding to domain_block_number.

domain_block_extrinsic_root: <T as Config>::DomainHash

Extrinsic root field of the header of domain block referenced by this ER.

parent_domain_block_receipt_hash: <T as Config>::DomainHash

The hash of the ER for the last domain block.

consensus_block_number: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

A pointer to the consensus block index which contains all of the bundles that were used to derive and order all extrinsics executed by the current domain block for this ER.

consensus_block_hash: <T as Config>::Hash

The block hash corresponding to consensus_block_number.

inboxed_bundles: Vec<InboxedBundle<<T as Config>::DomainHash>, Global>

All the bundles that being included in the consensus block.

final_state_root: <T as Config>::DomainHash

The final state root for the current domain block reflected by this ER.

Used for verifying storage proofs for domains.

execution_trace: Vec<<T as Config>::DomainHash, Global>

List of storage roots collected during the domain block execution.

execution_trace_root: H256

The Merkle root of the execution trace for the current domain block.

Used for verifying fraud proofs.

total_rewards: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

All SSC rewards for this ER to be shared across operators.