Crate pallet_domains

Pallet Domains

Modules

block_tree

Domain block tree

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

domain_registry

Domain registry for domains

extensions

Extensions for unsigned general extrinsics

migration_v2_to_v3

Migration module for pallet-domains EVM and AutoId domain config storage, versions 2 to 3.

runtime_registry

Runtime registry for domains

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_domains

Structs

EnsureDomainOrigin

Ensure the domain origin.

GenesisConfig

Can be used to configure the genesis state of this pallet.

OperatorConfig

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

StartingEVMChainId

Starting EVM chain ID for evm runtimes.

TxRangeState

Per-domain state for tx range calculation.

Enums

BundleError

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

FraudProofError

RawOrigin

Custom origin for validated unsigned extrinsics.

SlashedReason

Reason for slashing an operator

Constants

MAX_NOMINATORS_TO_SLASH

Maximum number of nominators to slash within a give operator at a time.

Traits

BlockSlot

Config

Configuration trait of this pallet.

HoldIdentifier

Functions

calculate_tx_range

Calculates the new tx range based on the bundles produced during the interval.

Type Aliases

BlockInherentExtrinsicData

Combined fraud proof data for the InvalidInherentExtrinsic fraud proof

BlockTreeNodeFor

ConsensusBlockHash

Map of consensus block hashes.

DomainBlockNumberFor

DomainHashingFor

DomainRuntimeUpgrades

Temporary storage to keep track of domain runtime upgrades which happened in the parent block. Cleared in the current block’s initialization.

DomainSudoCalls

Temporary storage to hold the sudo calls meant for domains.

EvmDomainContractCreationAllowedByCalls

Temporary storage to hold the “set contract creation allowed by” calls meant for EVM Domains.

ExecutionInbox

A set of BundleDigest from all bundles that successfully submitted to the consensus block, these bundles will be used to construct the domain block and ExecutionInbox is used to:

ExecutionReceiptOf

FraudProofFor

FrozenDomains

Storage that hold a list of all frozen domains.

LatestConfirmedDomainExecutionReceipt

Storage to hold all the domain’s latest confirmed block.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OpaqueBundleOf

OperatorEpochSharePrice

Share price for the operator pool at the end of Domain epoch.

Origin

ReceiptHashFor

RuntimeRegistry

Storage type is [StorageMap] with key type RuntimeId and value type RuntimeObject < BlockNumberFor < T > , T :: Hash >.

SingletonReceiptOf

SuccessfulBundles

Bundles submitted successfully in current block.