Crate pallet_domains

Pallet Domains

Re-exports

pub use staking::OperatorConfig;

Modules

block_tree

Domain block tree

bundle_storage_fund

Bundle storage fund

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

domain_registry

Domain registry for domains

extensions

Extensions for unsigned general extrinsics

fuzz

migrations

Migration module for Domains

runtime_registry

Runtime registry for domains

staking

Staking for domains

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_domains

Structs

EnsureDomainOrigin

Ensure the domain origin.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

TxRangeState

Per-domain state for tx range calculation.

Enums

BundleError

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

FraudProofError

RawOrigin

Custom origin for validated unsigned extrinsics.

SlashedReason

Reason for slashing an operator

Constants

MAX_NOMINATORS_TO_SLASH

Maximum number of nominators to slash within a give operator at a time.

Traits

BlockSlot

Config

Configuration trait of this pallet.

HoldIdentifier

WeightInfo

Weight functions needed for pallet_domains.

Functions

calculate_tx_range

Calculates the new tx range based on the bundles produced during the interval.

Type Aliases

BlockInherentExtrinsicData

Combined fraud proof data for the InvalidInherentExtrinsic fraud proof

BlockTreeNodeFor

ConsensusBlockHash

Map of consensus block hashes.

DeactivatedOperators

Storage for operators who were de-activated during this epoch. Will be cleared once epoch is transitioned.

DeregisteredOperators

Storage for operators who de-registered in the current epoch. Will be cleared once epoch is transitioned.

DomainBlockNumberFor

DomainChainRewards

Storage for chain rewards specific to each domain. These rewards to equally distributed to active operators during epoch migration.

DomainGenesisBlockExecutionReceipt

Storage to hold all the domain’s genesis execution receipt.

DomainHashingFor

DomainRuntimeUpgrades

Temporary storage to keep track of domain runtime upgrades which happened in the parent block. Cleared in the current block’s initialization.

DomainSudoCalls

Temporary storage to hold the sudo calls meant for domains.

EpochStartSlot

Stores the slot at which a new epoch has started.

EvmChainIds

Stored the occupied evm chain id against a domain_id.

EvmDomainContractCreationAllowedByCalls

Temporary storage to hold the “set contract creation allowed by” calls meant for EVM Domains.

ExecutionInbox

A set of BundleDigest from all bundles that successfully submitted to the consensus block, these bundles will be used to construct the domain block and ExecutionInbox is used to:

ExecutionReceiptOf

ExecutionReceiptRefOf

FraudProofFor

FrozenDomains

Storage that hold a list of all frozen domains.

InvalidBundleAuthors

Storage for operators who are marked as invalid bundle authors in the current epoch. Will be cleared once epoch is transitioned.

LatestConfirmedDomainExecutionReceipt

Storage to hold all the domain’s latest confirmed block.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OpaqueBundleOf

OperatorBundleCountInEpoch

Stores the number of bundles each operator submitted in a given epoch. Storage is cleared at the end of epoch.

OperatorEpochSharePrice

Share price for the operator pool at the end of Domain epoch.

Origin

PreviousBundleAndExecutionReceiptVersions

Storage that hold a previous versions of Bundle and Execution Receipt. Unfortunately, it adds a new item for every runtime upgrade if the versions change between runtime upgrades. If the versions does not change, then same version is set with higher block number. Pruning this storage is not quiet straight forward since each domain may submit an ER with a gap as well and also introduces the loop to find the correct block number.

ReceiptHashFor

RuntimeRegistry

Storage type is [StorageMap] with key type RuntimeId and value type RuntimeObject < BlockNumberFor < T > , T :: Hash >.

SingletonReceiptOf

SuccessfulBundles

Bundles submitted successfully in current block.