Crate pallet_domains

Pallet Domains

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

block_tree

Domain block tree

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

domain_registry

Domain registry for domains

runtime_registry

Runtime registry for domains

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_domains

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

TxRangeState

Per-domain state for tx range calculation.

Enums

BundleError

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

FreezeIdentifier

Functions

calculate_tx_range

Calculates the new tx range based on the bundles produced during the interval.

Type Definitions

ExecutionInbox

A set of bundle_extrinsics_root from all bundles that successfully submitted to the consensus block, these extrinsics will be used to construct the domain block and ExecutionInbox is used to ensure subsequent ERs of that domain block include all pre-validated extrinsic bundles.

ExecutionReceiptOf

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OpaqueBundleOf