Module pallet_domains::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

deregister_operator

Warning: Doc-Only

force_staking_epoch_transition

Force staking epoch transition for a given domain

instantiate_domain

Warning: Doc-Only

nominate_operator

Warning: Doc-Only

register_domain_runtime

Warning: Doc-Only

register_operator

Warning: Doc-Only

set_permissioned_action_allowed_by

Update permissioned action allowed by storage by Sudo.

submit_bundle

Warning: Doc-Only

submit_fraud_proof

Warning: Doc-Only

unlock_funds

Unlocks the first withdrawal given the unlocking period is complete. Even if rest of the withdrawals are out of unlocking period, nominator should call this extrinsic to unlock each withdrawal

unlock_operator

Unlocks the operator given the unlocking period is complete. Anyone can initiate the operator unlock.

update_domain_operator_allow_list

Extrinsic to update domain’s operator allow list. Note:

upgrade_domain_runtime

Warning: Doc-Only

withdraw_stake

Warning: Doc-Only