pallet_domains

Module staking

Staking for domains

Structs

DomainEpoch

Unique epoch identifier across all domains. A combination of Domain and its epoch.

NewDeposit

Operator

Type that represents an operator details.

OperatorConfig

OperatorDeregisteredInfo

SharePrice

A share price is parts per billion of shares/ai3. Note: Shares must always be equal to or lower than ai3, and both shares and ai3 can’t be zero.

StakingSummary

Enums

Error

OperatorStatus

Type that represents an operator status.

Functions

do_register_operator