pallet_domains

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccumulatedTreasuryFunds

Accumulate treasury funds temporarily until the funds are above Existential deposit. We do this to ensure minting small amounts into treasury would not fail.

BlockInherentExtrinsicData

Combined fraud proof data for the InvalidInherentExtrinsic fraud proof

BlockTree

The domain block tree, map (domain_id, domain_block_number) to the hash of ER, which can be used get the block tree node in BlockTreeNodes

BlockTreeNodes

Mapping of block tree node hash to the node, each node represent a domain block

ConsensusBlockHash

Map of consensus block hashes.

DeactivatedOperators

Storage for operators who were de-activated during this epoch. Will be cleared once epoch is transitioned.

DepositOnHold

The amount of balance the nominator hold for a given operator

Deposits

List of all deposits for given Operator.

DeregisteredOperators

Storage for operators who de-registered in the current epoch. Will be cleared once epoch is transitioned.

DomainChainRewards

Storage for chain rewards specific to each domain. These rewards to equally distributed to active operators during epoch migration.

DomainGenesisBlockExecutionReceipt

Storage to hold all the domain’s genesis execution receipt.

DomainRegistry

The domain registry

DomainRuntimeUpgradeRecords

Storage used to keep track of which consensus block each domain runtime upgrade happens in.

DomainRuntimeUpgrades

Temporary storage to keep track of domain runtime upgrades which happened in the parent block. Cleared in the current block’s initialization.

DomainStakingSummary

Warning: Doc-Only

DomainSudoCalls

Temporary storage to hold the sudo calls meant for domains.

DomainTxRangeState

Warning: Doc-Only

EpochStartSlot

Stores the slot at which a new epoch has started.

EvmChainIds

Stored the occupied evm chain id against a domain_id.

EvmDomainContractCreationAllowedByCalls

Temporary storage to hold the “set contract creation allowed by” calls meant for EVM Domains.

ExecutionInbox

A set of BundleDigest from all bundles that successfully submitted to the consensus block, these bundles will be used to construct the domain block and ExecutionInbox is used to:

FrozenDomains

Storage that hold a list of all frozen domains.

HeadDomainNumber

The block number of the best domain block, increase by one when the first bundle of the domain is successfully submitted to current consensus block, which mean a new domain block with this block number will be produce. Used as a pointer in ExecutionInbox to identify the current under building domain block, also used as a mapping of consensus block number to domain block number.

HeadReceiptNumber

The head receipt number of each domain

InboxedBundleAuthor

A mapping of bundle_header_hash -> bundle_author for all the successfully submitted bundles of the last BlockTreePruningDepth domain blocks. Used to verify the invalid bundle fraud proof and slash malicious operator who have submitted invalid bundle.

InvalidBundleAuthors

Storage for operators who are marked as invalid bundle authors in the current epoch. Will be cleared once epoch is transitioned.

LastEpochStakingDistribution

A temporary storage to hold any previous epoch details for a given domain if the epoch transitioned in this block so that all the submitted bundles within this block are verified. TODO: The storage is cleared on block finalization that means this storage is already cleared when verifying the submit_bundle extrinsic and not used at all

LatestConfirmedDomainExecutionReceipt

Storage to hold all the domain’s latest confirmed block.

LatestSubmittedER

The latest ER submitted by the operator for a given domain. It is used to determine if the operator has submitted bad ER and is pending to slash.

NewAddedHeadReceipt

The hash of the new head receipt added in the current consensus block

NextDomainId

Stores the next domain id.

NextOperatorId

Warning: Doc-Only

NextRuntimeId

Stores the next runtime id.

OperatorBundleCountInEpoch

Stores the number of bundles each operator submitted in a given epoch. Storage is cleared at the end of epoch.

OperatorBundleSlot

The set of slot of the bundle submitted by an operator in the current block, cleared at the next block initialization

OperatorEpochSharePrice

Share price for the operator pool at the end of Domain epoch.

OperatorHighestSlot

The highest slot of the bundle submitted by an operator

OperatorIdOwner

Warning: Doc-Only

Operators

List of all registered operators and their configuration.

PendingSlashes

A list operators who were slashed during the current epoch associated with the domain. When the epoch for a given domain is complete, operator total stake is moved to treasury and then deleted.

PendingStakingOperationCount

The pending staking operation count of the current epoch, it should not larger than MaxPendingStakingOperation and will be resetted to 0 upon epoch transition.

PermissionedActionAllowedBy

Storage for PermissionedActions for domain instantiation and other permissioned calls.

PreviousBundleAndExecutionReceiptVersions

Storage that hold a previous versions of Bundle and Execution Receipt. Unfortunately, it adds a new item for every runtime upgrade if the versions change between runtime upgrades. If the versions does not change, then same version is set with higher block number. Pruning this storage is not quiet straight forward since each domain may submit an ER with a gap as well and also introduces the loop to find the correct block number.

RuntimeRegistry

Warning: Doc-Only

ScheduledRuntimeUpgrades

Warning: Doc-Only

SuccessfulBundles

Bundles submitted successfully in current block.

Withdrawals

List of all withdrawals for a given operator.