Module pallet_domains::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

DomainRegistry

The domain registry

DomainStakingSummary

Warning: Doc-Only

DomainTxRangeState

Warning: Doc-Only

NextDomainId

Stores the next domain id.

NextOperatorId

Warning: Doc-Only

NextRuntimeId

Stores the next runtime id.

Nominators

Warning: Doc-Only

OperatorIdOwner

Warning: Doc-Only

Operators

Warning: Doc-Only

PendingDeposits

Warning: Doc-Only

PendingOperatorDeregistrations

Operators who chose to deregister from a domain. Stored here temporarily until domain epoch is complete.

PendingOperatorSwitches

Temporary hold of all the operators who decided to switch to another domain. Once epoch is complete, these operators are added to new domains under next_operators.

PendingWithdrawals

Warning: Doc-Only

RuntimeRegistry

Warning: Doc-Only

ScheduledRuntimeUpgrades

Warning: Doc-Only

SuccessfulBundles

Bundles submitted successfully in current block.