Module pallet_domains::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

BlockTree

The domain block tree, map (domain_id, domain_block_number) to the hash of a domain blocks, which can be used get the domain block in DomainBlocks

DomainBlocks

Mapping of domain block hash to domain block

DomainRegistry

The domain registry

DomainStakingSummary

Warning: Doc-Only

DomainTxRangeState

Warning: Doc-Only

ExecutionInbox

A set of bundle_extrinsics_root from all bundles that successfully submitted to the consensus block, these extrinsics will be used to construct the domain block and ExecutionInbox is used to ensure subsequent ERs of that domain block include all pre-validated extrinsic bundles.

HeadDomainNumber

The block number of the best domain block, increase by one when the first bundle of the domain is successfully submitted to current consensus block, which mean a new domain block with this block number will be produce. Used as a pointer in ExecutionInbox to identify the current under building domain block, also used as a mapping of consensus block number to domain block number.

HeadReceiptNumber

The head receipt number of each domain

LastEpochStakingDistribution

A temporary storage to hold any previous epoch details for a given domain if the epoch transitioned in this block so that all the submitted bundles within this block are verified. The storage is cleared on block finalization.

NextDomainId

Stores the next domain id.

NextOperatorId

Warning: Doc-Only

NextRuntimeId

Stores the next runtime id.

Nominators

List of all current epoch’s nominators and their shares under a given operator,

OperatorIdOwner

Warning: Doc-Only

Operators

List of all registered operators and their configuration.

PendingDeposits

Deposits initiated a nominator under this operator. Will be stored temporarily until the current epoch is complete. Once, epoch is complete, these deposits are staked beginning next epoch.

PendingGenesisDomain

The genesis domian that scheduled to register at block #1, should be removed once https://github.com/paritytech/substrate/issues/14541 is resolved.

PendingNominatorUnlocks

All the pending unlocks for the nominators. We use this storage to fetch all the pending unlocks under a operator pool at the time of slashing.

PendingOperatorDeregistrations

Operators who chose to deregister from a domain. Stored here temporarily until domain epoch is complete.

PendingOperatorSwitches

Temporary hold of all the operators who decided to switch to another domain. Once epoch is complete, these operators are added to new domains under next_operators.

PendingOperatorUnlocks

Stores a list of operators who are unlocking in the coming blocks. The operator will be removed when the wait period is over or when the operator is slashed.

PendingSlashes

A list operators who were slashed during the current epoch associated with the domain. When the epoch for a given domain is complete, operator total stake is moved to treasury and then deleted.

PendingUnlocks

A list of operators that are either unregistering or one more of the nominators are withdrawing some staked funds.

PendingWithdrawals

Withdrawals initiated a nominator under this operator. Will be stored temporarily until the current epoch is complete. Once, epoch is complete, these will be moved to PendingNominatorUnlocks.

RuntimeRegistry

Warning: Doc-Only

ScheduledRuntimeUpgrades

Warning: Doc-Only

SuccessfulBundles

Bundles submitted successfully in current block.