Trait pallet_domains::Config

pub trait Config: Config<Hash: Into<H256>> {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type DomainHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaxEncodedLen + Into<H256> + From<H256>;
    type DomainHeader: HeaderT<Hash = Self::DomainHash>;
    type ConfirmationDepthK: Get<BlockNumberFor<Self>>;
    type DomainRuntimeUpgradeDelay: Get<BlockNumberFor<Self>>;
    type Currency: Mutate<Self::AccountId> + InspectHold<Self::AccountId> + MutateHold<Self::AccountId>;
    type Share: Parameter + Member + MaybeSerializeDeserialize + Debug + AtLeast32BitUnsigned + FullCodec + Copy + Default + TypeInfo + MaxEncodedLen + IsType<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type HoldIdentifier: HoldIdentifier<Self>;
    type BlockTreePruningDepth: Get<DomainBlockNumberFor<Self>>;
    type MaxDomainBlockSize: Get<u32>;
    type MaxDomainBlockWeight: Get<Weight>;
    type MaxBundlesPerBlock: Get<u32>;
    type MaxDomainNameLength: Get<u32>;
    type DomainInstantiationDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
    type InitialDomainTxRange: Get<u64>;
    type DomainTxRangeAdjustmentInterval: Get<u64>;
    type MinOperatorStake: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type MinNominatorStake: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type StakeWithdrawalLockingPeriod: Get<DomainBlockNumberFor<Self>>;
    type StakeEpochDuration: Get<DomainBlockNumberFor<Self>>;
    type TreasuryAccount: Get<Self::AccountId>;
    type MaxPendingStakingOperation: Get<u32>;
    type MaxNominators: Get<u32>;
    type Randomness: RandomnessT<Self::Hash, BlockNumberFor<Self>>;
    type SudoId: Get<Self::AccountId>;
    type PalletId: Get<PalletId>;
    type StorageFee: StorageFee<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type BlockSlot: BlockSlot;
    type DomainsTransfersTracker: DomainsTransfersTracker<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type MaxInitialDomainAccounts: Get<u32>;
    type MinInitialDomainAccountBalance: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type DomainHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaxEncodedLen + Into<H256> + From<H256>

Domain block hash type.

type DomainHeader: HeaderT<Hash = Self::DomainHash>

The domain header type.

type ConfirmationDepthK: Get<BlockNumberFor<Self>>

Same with pallet_subspace::Config::ConfirmationDepthK.

type DomainRuntimeUpgradeDelay: Get<BlockNumberFor<Self>>

Delay before a domain runtime is upgraded.

type Currency: Mutate<Self::AccountId> + InspectHold<Self::AccountId> + MutateHold<Self::AccountId>

Currency type used by the domains for staking and other currency related stuff.

type Share: Parameter + Member + MaybeSerializeDeserialize + Debug + AtLeast32BitUnsigned + FullCodec + Copy + Default + TypeInfo + MaxEncodedLen + IsType<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Type representing the shares in the staking protocol.

type HoldIdentifier: HoldIdentifier<Self>

A variation of the Identifier used for holding the funds used for staking and domains.

type BlockTreePruningDepth: Get<DomainBlockNumberFor<Self>>

The block tree pruning depth.

type MaxDomainBlockSize: Get<u32>

The maximum block size limit for all domain.

type MaxDomainBlockWeight: Get<Weight>

The maximum block weight limit for all domain.

type MaxBundlesPerBlock: Get<u32>

The maximum bundle per block limit for all domain.

type MaxDomainNameLength: Get<u32>

The maximum domain name length limit for all domain.

type DomainInstantiationDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

The amount of fund to be locked up for the domain instance creator.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type InitialDomainTxRange: Get<u64>

Initial domain tx range value.

type DomainTxRangeAdjustmentInterval: Get<u64>

Domain tx range is adjusted after every DomainTxRangeAdjustmentInterval blocks.

type MinOperatorStake: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Minimum operator stake required to become operator of a domain.

type MinNominatorStake: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Minimum nominator stake required to nominate and operator.

type StakeWithdrawalLockingPeriod: Get<DomainBlockNumberFor<Self>>

Minimum number of blocks after which any finalized withdrawals are released to nominators.

type StakeEpochDuration: Get<DomainBlockNumberFor<Self>>

Domain epoch transition interval

type TreasuryAccount: Get<Self::AccountId>

Treasury account.

type MaxPendingStakingOperation: Get<u32>

The maximum number of pending staking operation that can perform upon epoch transition.

type MaxNominators: Get<u32>

The maximum number of nominators for given operator.

type Randomness: RandomnessT<Self::Hash, BlockNumberFor<Self>>

Randomness source.

type SudoId: Get<Self::AccountId>

The sudo account id

type PalletId: Get<PalletId>

The pallet-domains’s pallet id.

type StorageFee: StorageFee<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Storage fee interface used to deal with bundle storage fee

type BlockSlot: BlockSlot

The block slot

type DomainsTransfersTracker: DomainsTransfersTracker<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Transfers tracker.

type MaxInitialDomainAccounts: Get<u32>

Upper limit for total initial accounts domains

type MinInitialDomainAccountBalance: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Minimum balance for each initial domain account

Object Safety

This trait is not object safe.