Trait pallet_domains::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type DomainNumber: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + AtLeast32BitUnsigned + Default + Bounded + Copy + Hash + FromStr + MaxEncodedLen + TypeInfo;
    type DomainHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaxEncodedLen + Into<H256> + From<H256>;
    type ConfirmationDepthK: Get<Self::BlockNumber>;
    type DomainRuntimeUpgradeDelay: Get<Self::BlockNumber>;
    type Currency: Mutate<Self::AccountId> + MutateFreeze<Self::AccountId> + InspectFreeze<Self::AccountId>;
    type Share: Parameter + Member + MaybeSerializeDeserialize + Debug + AtLeast32BitUnsigned + FullCodec + Copy + Default + TypeInfo + MaxEncodedLen + IsType<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type FreezeIdentifier: FreezeIdentifier<Self>;
    type BlockTreePruningDepth: Get<Self::DomainNumber>;
    type MaxDomainBlockSize: Get<u32>;
    type MaxDomainBlockWeight: Get<Weight>;
    type MaxBundlesPerBlock: Get<u32>;
    type MaxDomainNameLength: Get<u32>;
    type DomainInstantiationDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
    type InitialDomainTxRange: Get<u64>;
    type DomainTxRangeAdjustmentInterval: Get<u64>;
    type MinOperatorStake: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>;
    type StakeWithdrawalLockingPeriod: Get<Self::DomainNumber>;
    type StakeEpochDuration: Get<Self::DomainNumber>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type DomainNumber: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + AtLeast32BitUnsigned + Default + Bounded + Copy + Hash + FromStr + MaxEncodedLen + TypeInfo

Domain block number type.

type DomainHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaxEncodedLen + Into<H256> + From<H256>

Domain block hash type.

type ConfirmationDepthK: Get<Self::BlockNumber>

Same with pallet_subspace::Config::ConfirmationDepthK.

type DomainRuntimeUpgradeDelay: Get<Self::BlockNumber>

Delay before a domain runtime is upgraded.

type Currency: Mutate<Self::AccountId> + MutateFreeze<Self::AccountId> + InspectFreeze<Self::AccountId>

Currency type used by the domains for staking and other currency related stuff.

type Share: Parameter + Member + MaybeSerializeDeserialize + Debug + AtLeast32BitUnsigned + FullCodec + Copy + Default + TypeInfo + MaxEncodedLen + IsType<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Type representing the shares in the staking protocol.

type FreezeIdentifier: FreezeIdentifier<Self>

Identifier used for Freezing the funds used for staking.

type BlockTreePruningDepth: Get<Self::DomainNumber>

The block tree pruning depth, its value should <= BlockHashCount because we need the consensus block hash to verify execution receipt, which is used to construct the node of the block tree.

TODO: BlockTreePruningDepth <= BlockHashCount is not enough to guarantee the consensus block hash must exists while verifying receipt because the domain block is not mapping to the consensus block one by one, we need to either store the consensus block hash in runtime manually or store the consensus block hash in the client side and use host function to get them in runtime.

type MaxDomainBlockSize: Get<u32>

The maximum block size limit for all domain.

type MaxDomainBlockWeight: Get<Weight>

The maximum block weight limit for all domain.

type MaxBundlesPerBlock: Get<u32>

The maximum bundle per block limit for all domain.

type MaxDomainNameLength: Get<u32>

The maximum domain name length limit for all domain.

type DomainInstantiationDeposit: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

The amount of fund to be locked up for the domain instance creator.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type InitialDomainTxRange: Get<u64>

Initial domain tx range value.

type DomainTxRangeAdjustmentInterval: Get<u64>

Domain tx range is adjusted after every DomainTxRangeAdjustmentInterval blocks.

type MinOperatorStake: Get<<<Self as Config>::Currency as Inspect<<Self as Config>::AccountId>>::Balance>

Minimum operator stake required to become operator of a domain.

type StakeWithdrawalLockingPeriod: Get<Self::DomainNumber>

Minimum number of blocks after which any finalized withdrawals are released to nominators.

type StakeEpochDuration: Get<Self::DomainNumber>

Domain epoch transition interval